[Refactor] #229 HomeViewModel 구현 , Recyclerview 구현
